pyproject.toml
[tool.poetry]
name = "overcier"
version = "0.1.0"
description = ""
authors = ["rvorias <raphaelvorias@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
aiohttp = "^3.9.0"
discord-py = "^2.3.2"
python-dotenv = "^1.0.0"
rich = "^13.7.0"
asyncpg = "^0.29.0"
instructor = "^0.4.0"
openai = "^1.3.6"
pydantic = "^2.5.2"
rembg = "^2.0.56"
glif-client = "^0.1.1"


[tool.poetry.group.dev.dependencies]
ruff = "^0.1.6"
pytest = "^8.1.1"
pytest-asyncio = "^0.23.6"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff.lint]
select = [
    # pycodestyle
    "E",
    # Pyflakes
    "F",
    # pyupgrade
    "UP",
    # flake8-bugbear
    "B",
    # flake8-simplify
    "SIM",
    # isort
    "I",
]